+++ packages/ddp-client/client/client_convenience.js
@@ -50,6 +50,7 @@ Meteor.connection = DDP.connect(ddpUrl, {
   'call',
   'callAsync',
   'apply',
+  'applyAsync',
   'status',
   'reconnect',
   'disconnect'